refactor: Split SearchBar into base component and create useDebounce hook
