+++ apps/web/src/features/spots/ui/SpotList.tsx
@@ -4,8 +4,8 @@ import { LoadingSpinner } from '@/fsd/shared';
 import { Box } from '@jung/design-system/components';
 import { Suspense, useState } from 'react';
 import { IoGridOutline, IoMapOutline } from 'react-icons/io5';
+import { SearchBar } from '../../../shared/ui/SearchBar';
 import { CategoryFilter } from './CategoryFilter';
-import { SearchBar } from './SearchBar';
 import { SpotContent } from './SpotContent';
 import * as styles from './SpotList.css';
 import SpotListSkeleton from './SpotListSkeleton';
