+++ apps/web/src/shared/ui/SearchBar.css.ts
