+++ apps/web/src/shared/ui/index.ts
@@ -9,3 +9,4 @@ export { AnimatedLine } from './AnimatedLine';
 export { Modal } from './Modal';
 export { ErrorBoundary } from './ErrorBoundary';
 export { MotionCard } from './MotionCard';
+export { SearchBar } from './SearchBar';
