+++ packages/shared/hooks/index.ts
@@ -2,3 +2,4 @@ export { useViewportSize } from './useViewportSize';
 export { useBeforeMatch } from './useBeforeMatch';
 export { useCollapsibleHeight } from './useCollapsibleHeight';
 export { usePortal } from './usePortal';
+export { useDebounce } from './useDebounce';
